Move plots and data into their own subdirectories
