+++ .gitignore
@@ -7,3 +7,4 @@
 */node_modules
 */htmlcov
 */__pycache__
+docs
