+++ frontend/src/data/boxplot.tsx
@@ -1,4 +1,4 @@
-import * as utils from '../../src/utils'
+import * as utils from '../utils'
 
 interface Row {
   cell: string
