+++ frontend/src/data/form.tsx
@@ -1,4 +1,4 @@
-import {Conf} from '../../src/Form'
+import {Conf} from '../Form'
 
 export const form_test_conf: Conf = {
   cells: ['CD4', 'CD4_Treg', 'CD8', 'CD8_Treg'],
