+++ frontend/src/data/splash.tsx
